+++ android/src/main/gen/cn/jystudio/bluetooth/Manifest.java
@@ -1,6 +1,6 @@
 /*___Generated_by_IDEA___*/
 
-package cn.jystudio.bluetooth;
+package com.dsreactnativebluetoothprinter;
 
 /* This stub is only used by the IDE. It is NOT the Manifest class actually packed into the APK */
 public final class Manifest {
