+++ android/src/main/java/com/dsreactnativebluetoothprinter/BluetoothServiceStateObserver.java
@@ -1,8 +1,8 @@
-package cn.jystudio.bluetooth;
+package com.dsreactnativebluetoothprinter;
 
 import java.util.Map;
 
- 
+
 public interface BluetoothServiceStateObserver {
     void onBluetoothServiceStateChanged(int state, Map<String,Object> boundle);
 }
