+++ android/src/main/java/com/dsreactnativebluetoothprinter/RNBluetoothEscposPrinterPackage.java
@@ -1,12 +1,12 @@
 
-package cn.jystudio.bluetooth;
+package com.dsreactnativebluetoothprinter;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import cn.jystudio.bluetooth.escpos.RNBluetoothEscposPrinterModule;
-import cn.jystudio.bluetooth.tsc.RNBluetoothTscPrinterModule;
+import com.dsreactnativebluetoothprinter.escpos.RNBluetoothEscposPrinterModule;
+import com.dsreactnativebluetoothprinter.tsc.RNBluetoothTscPrinterModule;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
